+++ eslint.config.js
@@ -1,7 +1,7 @@
 import globals from 'globals';
 import pluginJs from '@eslint/js';
 import tseslint from 'typescript-eslint';
-import pluginVue, { rules } from 'eslint-plugin-vue';
+import pluginVue from 'eslint-plugin-vue';
 import markdown from '@eslint/markdown';
 
 export default [
@@ -14,5 +14,11 @@ export default [
   {
     files: ['**/*.vue'],
     languageOptions: { parserOptions: { parser: tseslint.parser } },
+  },
+  {
+    rules: {
+      'vue/multi-word-component-names': 'off',
+      '@typescript-eslint/no-explicit-any': 'off',
+    },
   },
 ];