+++ src/components/form/Form.vue
@@ -4,9 +4,6 @@
       'nt-form': true,
       'nt-form-inline': inline,
     }"
-    :style="{
-      '--nt-form-label-width': labelWidth ? labelWidth : undefined,
-    }"
     @submit="handleSubmit"
   >
     <slot></slot>
@@ -29,10 +26,12 @@ const props = withDefaults(
     /** 是否行内表单 */
     inline?: boolean;
     disabled?: boolean;
+    labelPosition?: 'left' | 'right' | 'top';
   }>(),
   {
     inline: false,
     disabled: false,
+    labelPosition: 'right',
   },
 );
 
@@ -61,7 +60,7 @@ if (props.model != null && validator != null) {
   const watcher = [];
   for (const key in props.model) {
     keys.push(key);
-    //@ts-ignore
+    //@ts-expect-error: Unreachable code error
     watcher.push(() => props.model[key]);
   }
   if (watcher.length > 0) {
@@ -92,6 +91,8 @@ provide(formContext, {
   errors,
   /** 必填字段列表 */
   requiredKeys,
+  labelWidth: () => props.labelWidth,
+  labelPosition: () => props.labelPosition,
 });
 
 provide(formDisabledContext, () => props.disabled);
@@ -99,6 +100,17 @@ provide(formDisabledContext, () => props.disabled);
 function handleSubmit(e: Event) {
   e.preventDefault();
   if (validator != null) {
+    validate();
+  } else {
+    emits('submit');
+  }
+}
+
+/**
+ * 校验全部表单数据
+ */
+function validate() {
+  if (validator != null && props.model != null) {
     validator
       .validate(props.model)
       .then(() => {
@@ -110,8 +122,48 @@ function handleSubmit(e: Event) {
           [err.key]: err.message,
         };
       });
-  } else {
-    emits('submit');
   }
 }
+
+/**
+ * 校验部分表单数据
+ * @param field
+ */
+function validateField(field: string | string[]) {
+  if (validator != null && props.model != null) {
+    const tacks: Promise<{
+      key: string;
+      value: any;
+    }>[] = [];
+    if (Array.isArray(field)) {
+      for (let i = 0, len = field.length; i < len; i++) {
+        tacks.push(
+          validator.validateKey(field[i], props.model[field[i]], props.model),
+        );
+      }
+    } else {
+      tacks.push(validator.validateKey(field, props.model[field], props.model));
+    }
+    Promise.allSettled(tacks).then((results) => {
+      for (let i = 0, len = results.length; i < len; i++) {
+        const result = results[i];
+        if (result.status === 'rejected') {
+          errors.value[result.reason.key] = result.reason.message;
+        } else {
+          errors.value[result.value.key] = undefined;
+        }
+      }
+    });
+  }
+}
+
+function clearValidate() {
+  errors.value = {};
+}
+
+defineExpose({
+  validate,
+  validateField,
+  clearValidate,
+});
 </script>