+++ src/components/input/Input.vue
@@ -27,12 +27,15 @@ const props = withDefaults(
     parser?: (value: string) => string;
     modelValue?: string | number;
     disabled?: boolean;
+    /** number, integer */
+    allowInput?: string;
   }>(),
   {
     htmlType: 'text',
     placeholder: '',
     autosize: false,
     disabled: undefined,
+    allowInput: undefined,
   },
 );
 
@@ -46,10 +49,49 @@ function focus() {
   }
 }
 
+function numberInputParse(
+  value: string,
+  config: { integer: boolean; negative: boolean; precition: number },
+) {
+  let val = value;
+  let negative = config.negative ? '-?' : '';
+  if (config.integer) {
+    const match = val.match(new RegExp(`^(${negative}\\d*)`));
+    if (match != null) {
+      return match[1];
+    }
+    return val.substring(0, val.length - 1);
+  }
+  const match = val.match(
+    new RegExp(
+      `(${negative}\\d+\\.\\d{0,${config.precition}})|(${negative}\\d*)`,
+    ),
+  );
+  if (match != null) {
+    val = match[1] || match[2];
+  } else {
+    val = '';
+  }
+  return match != null ? match[1] || match[2] : '';
+}
+
 function handleInput(e: Event) {
   const $target = e.target as HTMLInputElement;
   let value = $target.value;
   emits('input', e);
+  if (props.allowInput != null) {
+    let dotIndex = props.allowInput.indexOf('.');
+    let precition =
+      dotIndex === -1
+        ? dotIndex
+        : parseInt(props.allowInput.substring(dotIndex + 1));
+    value = numberInputParse(value, {
+      integer: props.allowInput.includes('integer'),
+      negative: props.allowInput.startsWith('-'),
+      precition: precition,
+    });
+    $target.value = String(value);
+  }
   if (props.parser != null) {
     value = props.parser(value) as string;
     $target.value = String(value);