+++ src/index.ts
@@ -3,7 +3,6 @@ export type { ColumnOption } from './components/table/Table.vue';
 export { default as BaseIcon } from './components/icon/Base.vue';
 export { default as MaskCloseIcon } from './components/icon/MaskClose.vue';
 export { default as SearchIcon } from './components/icon/Search.vue';
-export { default as IconFont } from './components/icon/IconFont.vue';
 export { default as LoadingIcon } from './components/icon/Loading.vue';
 export { default as InfoIcon } from './components/icon/Info.vue';
 export { default as WarnIcon } from './components/icon/Warn.vue';
@@ -25,6 +24,7 @@ export { default as DArrowRight } from './components/icon/DArrowRight.vue';
 export { default as SunIcon } from './components/icon/Sun.vue';
 export { default as MoonIcon } from './components/icon/Moon.vue';
 export { default as ThemeDefaultIcon } from './components/icon/ThemeDefault.vue';
+export { default as FontIcon } from './components/icon/IconFont.vue';
 
 export { default as Input } from './components/input/Input.vue';
 export { default as MdInput } from './components/input/MdInput.vue';
@@ -60,6 +60,7 @@ export { default as Polygon } from './components/Polygon.vue';
 export { default as Shadow } from './components/Shadow.vue';
 
 export { default as useFormReset } from './hooks/useFormReset';
+
 export { default as Collapse } from './components/collapse/Collapse.vue';
 export { default as CollapseItem } from './components/collapse/CollapseItem.vue';
 export { default as Progress } from './components/Progress.vue';
@@ -70,6 +71,8 @@ export { default as List } from './components/List.vue';
 
 export { default as Clickoutside } from './directives/clickoutside';
 export { default as Loading } from './directives/loading';
+export { default as BorderLoading } from './directives/borderLoading';
+
 export { default as PageHeader } from './components/PageHeader.vue';
 export { default as Tag } from './components/Tag.vue';
 export { default as Pagination } from './components/Pagination.vue';